[IMP] im_livechat: introduce chatbot scripts
PURPOSE

This commit introduces a chatbot operator that works based on a user-defined
script with various steps.

SPECS

A im_livechat.chatbot.script can be defined on a livechat rule.
When a end-user reaches a website page that matches the rule, the chat window
opens and the script of the bot starts iterating through its steps.

A script is defined with several steps that can currently be one of the
following types:
- text
  A simple text step where the bot posts a message without expecting an answer
  e.g: "Hello! I'm a friendly robot!"
- question_selection
  The bot will ask a question and suggest answers, the end-user will have to
  click on the answer he chooses
  e.g: "How can I help you?
  -> Create a Ticket
  -> Create a Lead
  -> Speak with a human"

The chatbot code is currently directly integrated with the existing livechat
Javascript code.
It defines extra conditions and layout elements to be able to automate the
conversation and register user answers.

Task-2030386

Co-authored-by: Patrick Hoste <pko@odoo.com>
Co-authored-by: Aurélien Warnon <awa@odoo.com>