[FIX] web: fix the color_picker widget alignment issue
Currently, the color picker widget is resized vertically instead of horizontally
if there is no record in planning resources.

In this commit we have fixed the issue, when there is no record in
the planning_resources list view, the color_picker widget gets 72px width by
default, so we have added a condition that if there is a color_picker widget
in a field, we will calculate the width based on the total fields in the view.

task-2720976
closes: #83727