[IMP] portal: prevent duplicate views of portal wizard
The portal's granting access / revoking access / reinvite actions
return the wizard modal opening action in order to keep it open
and allow the user to keep browsing and managing portal access.

However, this can return the modal opening several times and execute
them all, generating several (unsynchronized) modal view instances
that, the user sees and has to close one at a time.

Therefore the form view of the wizard receives a js class, that
only allows one of these actions at a time. Clicking on other ones
while the action is ongoing will not do anything since buttons are
disabled for the time the action is being done.

Task-2613780