[FIX] web_editor: resolve ajax request of cleanForSave
Upon save in website or in mass_mailing, if a `cleanForSave` of a
`SnippetEditor` or a `SnippetOption` wait for the end of an
`this._rpc()` request AND the `SnippetEditor` or the `SnippetEditor`
of the `SnippetOption` has a `SnippetEditor` ancestor, the
`cleanForSave` will never resolve.

Here is the scenario:

- `SnippetEditor B` has `SnippetEditor A` as it's odoo widget parent

In `SnippetMenu` `_destroyEditors`:
- `SnippetEditor A` and `all it's SnippetOption` call `cleanForSave`.
- `SnippetEditor B` and `all it's SnippetOption` call `cleanForSave`.
  One of the promise of `SnippetEditor B` includes the result of
  `rpc()` in `ajax_service.js`.

- `SnippetEditor A` and `all it's SnippetOption`
  `cleanForSave promises` resolve (but not `SnippetEditor B`).
- `SnippetEditor A` get's destroyed (`snippetEditor.destroy()`). As
`SnippetEditor B` is a `Widget` child of `SnippetEditor A`:
- `SnippetEditor B` get's destroyed (here is where the problem start).

... in the mean time ...

The ajax request of a promise of `SnippetEditor B` ends. The promise
of the `rpc()` will never finish because `SnippetEditor B` is
destroyed and the condition in `rpc()` is:
```js
if (!target.isDestroyed()) {
    resolve(result);
}
```

Because the promise will never finish for `_destroyEditors`, the save
action will never fully complete.