[REF] web: update rainbow man and revamp scss
Prior this commit the rainbow man was built with many separate SVGs and
a lot of raw SCSS.
This commit refactors the rainbow man in order to use less raw SCSS and
more SVG attributes instead.

It also improves the overall animation with a little rotation of the
rainbow man's head during the fade in animation.

task-2742831