[IMP] sale_timesheet: allow clicking on Timesheet sections
Before this commit, the user can see the profitability of his project
in the right side panel shown in the project update of this project.
However, if he wants to see the timesheets related to a section in the
project profitability he should search in this project.

This commit adds a link to allow clicking on a section about Timesheets
in costs and revenues section of the project profitability. This link
will redirect to the views containing the timesheets related to the
section clicked.

task-2558889