[FIX] mail: limit reply-to name part
Due to a python limitation we may have issues if reply-to of a sent email
contains more than 78 characters. Even if this is technically ok with the
RFC python seems to incorrectly handle it (please refer to [1] for more
details and discussions).

Until this is finally sorted out we decided to avoid issues by shortening
reply-to. Heuristic is now
  * either ``"CompanyName RecordName" <email@domain.com>`` is under 78 and we
    keep it;
  * otherwise we try by stripping out CompanyName as this information is not
    crucial, keeping ``"RecordName" <email@domain.com>``;
  * if this value is > 78 we strip record name extra characters, ending with
    something like ``"TooLong [...]"" <email@domain.com>``;
  * if email address itself is too long it is kept as it is without formating;

Task-2602862
OPW-2733513

[1] See https://bugs.python.org/issue44637