[IMP] point_of_sale: usability of cash-move-popup
Before, the cash move popup was overlapping with
the virtual keyboard in some devices. In this
commit we address the above issue.

We also make the popup more usable by changing
the order of the confirm and cancel buttons.

Finally, we make it simpler for the user by
making it not draggable.

task-2669747