[FIX] product,stock,stock_picking_batch: barcode type use on repport
Create 2 products with respectively barcodes 07010022 and 07010023
print labels for these products.

Result:
- barcodes on the labels are 07010022 and 07010023 (OK)
- barcode image on the labels are the same, when you read it with
a scanner, it says  07010026 -> not consistent

For the moment, the barcode type is selected according only to its
length.
With this commit, the preferred type for a barcode is "EAN8" or
"EAN13" but if the checksum doesn't correspond, it's using "code128".

opw-2719022