[REF] *sale*: clean inheritance/setup of sale report
* remove unused delay column (there was no field using the value server 
side)
* remove unused joins in the FROM part of the table definition (join on 
pricelists)
* fix broken inheritance API

pos_sale inheritance adds a null value for fields not available added in 
sub modules (website, invoice_status, ...).  But this was only working 
because the default value of the fields parameter of the _query method 
was overridden to a dict in all overrides, filling the record while 
climbing the inheritance chain.
This was clearly too magic, using information provided higher in the 
inheritance chain, assuming you were calling the super before generating 
the pos part of the query.

This commit cleans the methods and API of the model, providing clean 
hooks and reducing the mess of values given in overrides of _query.