[FIX] web: fix texts are overlapping
currently,
in project task view, in 'Blocked by' sub tab when there is no line and line is
added 'view task' button overlap with content of  previous cell.this is hppening
because the over flow of button.

after this commit,
'view task' button will not overlap with content of  previous cell. to solve
this issue set over flow=hidden.

closes-84325
task-2739646