[IMP] website_blog: align breadcrumbs and tags to post content sections
With the new layout introduced in the previous commit and the
suppression of the "Increase Readability" customization, the breadcrumbs
and tags did remain aligned with sections set as `o_container_small`.

This commit makes sure that the newly dropped blocks inside a blog post
are indeed set as `o_container_small` by default for the full cover
layout, and to `container` for the "Regular Cover" layout.
But in case this is changed, the breadcrumbs and tags are realigned with
the first text section (or the first section if there is no text
section within the blog post content).
These positions are re-evaluated both on page load and dynamically while
the blog post content is being edited.
Also aligns Comments with the content and makes Regular Cover's
Read Next span over the sidebar if it exists.

task-2449620