[IMP] pos_adyen: handle adyen payment request by push notification
Instead of eager polling of payment status of a payment request at
every 5.5 secs for some number of tries, we are now using the longpolling
bus service to receive in the POS UI the payment status in near real-time.

Basically, when adyen server notifies the server for the payment status,
the payment status will be automatically broadcasted. Open POS UI will
receive the notification and will handle it accordingly.

A simple workflow is also introduced to handle payment status that takes
a while to arrive:

After 1 min of valid payment request, the cashier is given an option to
either i) wait for another minute, ii) force done, or iii) retry the payment.
Note that forcing the payment to be done won't have the right
information about the adyen payment.