[FIX] account_edi_peppol: BE, DE, FR xml pass in ecosio and in Odoo
Belgium: the credit note and invoice xml now pass the test on ecosio for formats:
* OpenPEPPOL Invoice (3.13.0) (aka BIS Billing 3.0.12)
* OpenPEPPOL Credit Note (3.13.0) (aka BIS Billing 3.0.12)

Deutschland: same:
* XRechnung UBL Invoice 2.2.0
* XRechnung UBL CreditNote 2.2.0

France: same:
* UBL Invoice 2.1
* UBL Credit Note 2.1

Put all the tests in a new module. The xmls in /test_files makes it easier to check
that the tests are up-to-date: simply check each of these in ecosio.

The xml are embedded in a PDF by default. Also correct some bugs (zero division,
line.replace if line is None, typos, etc).

task-2628093