[FIX] web_editor, website: prevent deletion of some elements
This commit allows to target specific elements and prevent them from
being deleted from a page when editing it with the web_editor.

It adds a new property on the wysiwyg widget:
`unremovableElementsSelector` passed as an option to the class
constructor.
It will make sure that any elements matched by this selector won't
be removed either by OdooEditor or by the SnippetEditor.

task-2669592