[FIX] website, website_*: replace <b> tags by inline style
Since we changed the behavior of the text editor so that it applies
"bold" on the text and no longer "bolder" with the bold option. we have
replaced all the <b> tags of the editable content of website modules
with inline style ("font-weight: bold").

Indeed, with Bootstrap, the <b> tag applies "font-weight: bolder" to the
texts. What we no longer want and which is no longer possible to do with
the text editor.

task-2702644