[FIX] web_editor, *: fix image options visibility
*: website

Issue: from 15.0 onwards all image options are available on "/shop".

Details:

On v14.0, images options were removed for unsupported images (we had
a message: "Quality Options Unavailable" - see [1]) and images related
options on old editor toolbar were hidden too.

On v14.4 and after [2], images options from the toolbar (crop, width,...)
were moved to a separate 'ImageTools' option and as a consequence, they
don't respect the same visibility rules as 'ImageOptimize'. Other options
were added in XML and they have the same issue, they are visible as long
as the target is an 'img'.

This is a list of options that should be available for unsupported or
partly supported images:

[A]- for "t-field" images (shape & quality options are available only if
the image is supported for processing):

- Shape
- Filter
- Width (image width)
- Quality
- Transform (crop)

[B]- for external images or images unsupported for processing:

- Description
- Tooltip
- Transform (crop, transform)
- width (CSS width)
- Alignment
- Style
- Padding
- Animate

[C]- for illustrations:

- Same as [B]
- Dynamic colors

The goal of this commit is to fix this visibility issue on the JS code
of image options.

Remark: The '_computeWidgetVisibility()' method was duplicated on the
same 'WebsiteAnimate' option in [3], it's adapted in this commit too.

[1]: https://github.com/odoo/odoo/commit/84417a174bc2bb76730cc89783d3426a3a3b07d9
[2]: https://github.com/odoo/odoo/commit/d619c522205d964e58b8cfe2008079170ca41a73
[3]: https://github.com/odoo/odoo/commit/187acb938f70a2130d25fa76079221339c742f1e

task-2724946