[FIX] crm: use updated frequencies to compute probability on stage change
TO REPRODUCE:

- Create an empty db (no demo data for crm)
- create one lead
- set it won. Probability is set to 100% as expected
- change the stage back to a previous stage
- Proba is now 91,67% (1.1/1.2)

ISSUE:

This is because when changing to the previous stage, the frequency table is
updated AFTER the update of the probability value
- First the compute sets the new probability value using the table
- Then in the write, we have the table update

FIX:

If the stage / active value is to be changed in the write, recompute
probabilities. In that case, it will display 50% instead, since the table
is correctly updated and lost_count = won_count = 0.1.

Task-2673133