[FIX] web: fix strange behaviour in project-planning
before this commit,

in mobile view, in project>planning(gannt view) and click on 'plann existing' it
open kanban view  with quick create in dialog. but in planning it open kanban
with no content helper in dialog.

after this commit,

n mobile view, in project>planning(gannt view) and click on 'plann existing' it
open kanban view  without quick create in dialog.  to solve this issue checked
wether kanban view is open in dialog or not basedd on that and make
'group_creatable' to  'True' or 'False'.

no-content helper will not be there because it is intensional behaviour.
check out link:
https://github.com/odoo/odoo/commit/9445d2f04d0decc29366aedea819b091f287f5e5

closes-84325
task-2739646