[FIX] website: prevent pausing hover effect animations
Animations on hover effect content ends up in a paused state whenever they are
not displayed at the time the WebsiteAnimate widget is started. Since the
animation should be played each time the content is triggered, it is perfectly
reasonable to keep them in the running state at all times. Removing hover
content elements from the logic of the WebsiteANimate widget resolves the
issue.

task-2666891