[IMP] web: open document links in new tabs
With this commit, link in document will be open in new tab
so that Odoo document is not lost. For that we need to do
change in pdfjs library as we cannot pass this parameter
in iframe. Allowed parameters are 'page', 'zoom', 'nameddest',
'pagemode' and 'file'.

We have updated the value of 'externalLinkTarget' from 0 to 2
in 'getDefaultPreferences' function. '{NONE: 0, SELF: 1,
BLANK: 2, PARENT: 3, TOP: 4}' are the options for link target
and so we have to pass '2'(BLANK) as we need to open link in
new tab.

TaskId:2752002