[FIX] web_editor: block options panel until its changes are applied
Cherry-pick of #84688 - to be removed.

Before this commit, modifying an option in the options tab did only put
a blocking loader overlay on the page, but not on the options tab.
This made it possible to modify several options while a single change
was being applied.

After this commit when a change is done in the options tab an additional
blocking loader overlay is put on top of the options tab until de change
is fully applied.

task-2633169