[FIX] web_editor: can no longer break editor as Restricted Editor
When assigned the role of Restricted editor, there are many ways
to break the web_editor when trying to do actions that should be
unavailable. There is a traceback for the following scenarios:

- 	Drag and dropping snippet when Restricted Editor
- 	Clicking on product when Restricted Editor in /shop
- 	Clicking on product image on specific product page
- 	Clicking on user name (e.g. Marc Demo)
- 	Clicking on menu items or logo
- 	Clicking on a blog's image in /blog
- 	Clicking on a blog's image on specific blog page
- 	Clicking on calendar's image in /calendar

There is now a small notification indicating that the attempted action is
unauthorized and there is no longer a traceback which makes the editor
crash or freeze.