[WIP] website, web_editor: move website editor to the backend
This commit moves the edit button to the backend.

The goal is also to run the editor next to an iframe (as opposed to
inside the iframe like in mass_mailing).

What works:
 - Wysiwyg and SnippetMenu (and so on) start and load their UI
 - Dragging a snippet.

What doesn't work:
 - If you move the cursor too fast the snippet thumbnail stays behind
   if hover the iframe.
 - Dropping a snippet
 - Edit text
 - Saving the page

This is a huge WIP and hopefully future commits will make it all work :)