[FIX] mass_mailing: properly convert base64 images
The function that converts inline images to urls returned the modified
html as a byte string. As a result, the Mail Debug tab showed a
hexadecimal string rather than rendered html. This converts it back to
a regular utf-8 string so the format of the string coming out of the
function is the same as it was coming in.