[IMP] point_of_sale: visual improvments in clientListScreen
In this commit we add cursor lowlighting
for the client rows and the edit button.
We also add visual feedback whenever the
user clicks at a client row.

On top of that, we display the selected
client (if any) at the top of the screen
in order to ease editing and deselecting
workflow.

Finally, we change the text color
of the company name field to a different
shade of grey, to make it more distinct
from the client name field.

task-2669747