[FIX] website: prevent the deletion of the form's default values
Some form fields are dynamically filled. As the inputs that hold these
values do not always have a value property, their HTML value attribute
can be modified. Therefore the form option deletes the values so that
they do not become default values (see [1]). Unfortunately this system
is problematic and can remove default values set by the user. This
commit ensures that only auto-fillable fields and values from data-for
values will be removed by the option afterwards.

Test is added with this commit to verify that the default values are not
erased on later form edit.

Step to reproduce the fixed issue:
- Create a form, add a (required hidden) field with a default value
- Save, then enter edit mode again
- Add a new field and save again

The default values of the form are gone. Enter edit mode again to see
the hidden input, you will see there is no value.

[1]: odoo@b637a5e

task-2715201