[FIX] website_twitter: prevent empty scroller after drag and drop
The contents of the the twitter scroller snippet are provided by the public
widget. When the snippet is first added a loading tweets gif is shown. However,
currently the timeline is emptied again in the destroy method of the options
widget, resulting in an empty timeline whenever the options is restarted.

This happens, for example, after dragging and dropping the snippet to a new
location.

task-2735591