[IMP] website_crm_partner_assign: improve the kanban card view
Currently, if we have both customer and assigned partner names,
they are displayed in a single line separated by an arrow icon.
however, if we have long customer name, the arrow icon and
partner name are not visible on card.

This commit improves the display of names by making a few changes
    - breaking the line before the arrow icon.
    - customer and partner names now have a tooltip with their full names.
    - removing the arrow tooltip.

taskID: 2763930