[FIX] {sale,purchase}_stock: compute delivered/received qty
When using several UoMs, the delivered quantity on a POL can be bigger
than the reality

To reproduce the issue:
1. In Settings, enable "Units of Measure"
2. Create a new UoM U96:
    - Category: Unit
    - Type: Bigger
    - Bigger Ratio: 96
    - Rounding Precision: 1
3. Create a product P:
    - UoM: Unit
    - Purchase UoM: Unit
4. Create & Confirm a new PO with:
    - 2 U96 x P
5. Receive 144 units of P (with backorder)
    - Note that there is already an error: back to the PO, the received
quantity is 2 U96. This would mean that all P have been received, which
is not true
6. Process the backorder
7. Go back to the PO

Error: The qty received is now 3 U96. It's more than expected and than
reality

The current computation of the received quantity creates the confusion:
it rounds each SML and then sum the rounded values. So, in the above
case, from the first picking, we received 144 units, which means 1.5
U96. We round this value: 2 U96. Then, we do the same with the
backorder: 48 units = 0.5 U96 => 1 U96. Therefore: 2 U96 + 1 U96 = 3 U96

A similar issue occurs with the SO.

OPW-2748532